requirements.txt
cffi==1.15.0
click==8.1.2
Flask==2.1.1
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
protobuf==3.19.4
pycparser==2.21
PyJWT==2.4.0
six==1.16.0
Werkzeug==2.1.1
zipp==3.8.0